moved base data processor to its own file
