+++ test/data_processing/__init__.py
@@ -1,0 +1,1 @@
+from .base_data_processor_test import BaseDataProcessorTest
