+++ test/data_processing/test_nodes.py
@@ -14,7 +14,7 @@
 
 # pylint: disable=unbalanced-tuple-unpacking
 
-from test.fake_experiment import FakeExperiment, BaseDataProcessorTest
+from test.fake_experiment import FakeExperiment
 
 from typing import Any, List
 import numpy as np
@@ -26,6 +26,7 @@ from qiskit_experiments.experiment_data import ExperimentData
 from qiskit_experiments.data_processing.nodes import SVD, AverageData, MinMaxNormalize
 from qiskit_experiments.data_processing.data_processor import DataProcessor
 
+from . import BaseDataProcessorTest
 
 class TestAveraging(QiskitTestCase):
     """Test the averaging nodes."""
